requirements.txt
# PySide6 (Qt6)
PySide6>=6.6.0

# HTTP 요청
requests>=2.31.0

# HTML 파싱
beautifulsoup4>=4.12.0

# 이미지 처리 (아이콘)
Pillow>=10.0.0

# Firebase (필수)
pyrebase4>=4.7.0

# 메모리 로그 (선택)
# psutil>=5.9.0